backend/requirements.txt
Flask==2.1.0
numpy==1.21.0
opencv-python==4.5.3.56
flask-cors==3.0.10
scikit-image==0.18.3
Pillow==8.2.0
